requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
aiosqlite==0.20.0
python-dotenv==1.0.0
